actions: run tests as separate steps
There are no files directly in xt/ currently, but may be in the future.
Using xt/*.t will result in an error. The nullglob option can't be use
because the shell doesn't support it.

Instead, just name the directory and have a separate step for the
xt/author directory that works recursively.